requirements.txt
pydantic>=2.0,<3
pyyaml>=6.0
# ── connectors (install as needed) ──
# python-telegram-bot>=20.0
# openai>=1.0
# ── dev/test ──
pytest>=8.0
